LPD-75170 refactor: rename ai-hub-web to ai-hub-nexus-web
