+++ build.properties
@@ -292,7 +292,7 @@
     #    core/portal-web,\
     #    core/shielded-container-api,\
     #    core/shielded-container-impl,\
-    #    dxp/apps/ai-hub,\
+    #    dxp/apps/ai-hub-nexus,\
     #    dxp/apps/analytics,\
     #    dxp/apps/antivirus,\
     #    dxp/apps/commerce,\
