+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-web/node-scripts.config.js
