+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-web/package.json
@@ -1,0 +1,28 @@
+{
+	"dependencies": {
+		"@clayui/alert": "^3.159.0",
+		"@clayui/button": "^3.159.0",
+		"@clayui/core": "^3.159.0",
+		"@clayui/form": "^3.159.0",
+		"@clayui/icon": "^3.159.0",
+		"@clayui/layout": "^3.159.0",
+		"@clayui/link": "^3.159.0",
+		"@clayui/panel": "^3.159.0",
+		"@clayui/provider": "^3.159.0",
+		"@clayui/toolbar": "^3.159.0",
+		"@liferay/object-js-components-web": "*",
+		"frontend-js-components-web": "*",
+		"frontend-js-web": "*",
+		"react": "18.2.0",
+		"react-dom": "18.2.0"
+	},
+	"main": "js/index.ts",
+	"name": "@liferay/ai-hub-nexus-web",
+	"scripts": {
+		"build": "node-scripts build",
+		"checkFormat": "node-scripts format --check",
+		"format": "node-scripts format",
+		"test": "node-scripts test"
+	},
+	"version": "1.0.2"
+}
