+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-web/src/main/resources/META-INF/resources/view_agent_definitions.jsp
