--- modules/dxp/apps/ai-hub-nexus/ai-hub-web/src/main/java/com/liferay/ai/hub/web/internal/constants/AIHubFDSNames.java
@@ -1,16 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.ai.hub.web.internal.constants;
-
-/**
- * @author Feliphe Marinho
- */
-public class AIHubFDSNames {
-
-	public static final String AGENT_DEFINITIONS =
-		AIHubWebConstants.BUNDLE_SYMBOLIC_NAME + "-agentDefinitions";
-
-}
