+++ modules/dxp/apps/ai-hub-nexus/test.properties
@@ -7,6 +7,8 @@ modules.includes.required.test.batch.class.names.includes[modules-integration-po
     dxp/apps/ai-hub-nexus/**/*Test.java
 
 relevant.rule.names=ai-hub-nexus-java-rule
-test.batch.names[relevant][ai-hub-nexus-java-rule]=modules-integration-postgresql163
+
+test.batch.names[relevant][ai-hub-nexus-java-rule]=\
+    modules-integration-postgresql163
 
 testray.main.component.name=AI Hub Nexus
